requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu>=2.10.0
pillow==10.0.1
numpy>=1.21.0,<2.0.0
google-cloud-storage==2.10.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0